requirements.txt
pytest==8.3.2
h11==0.14.0
pip==23.2.1
attrs==22.2.0
typing_extensions==4.12.2
trio==0.24.0
outcome==1.3.0.post0
sniffio==1.3.1
sortedcontainers==2.4.0
cffi==1.17.0
idna==3.7
mysql-connector-python==9.0.0
Appium-Python-Client==1.3.0
selenium==3.141.0
urllib3==1.26.8
pluggy==1.5.0
iniconfig==2.0.0
certifi==2024.7.4
six==1.16.0
pycparser==2.22
websocket-client==1.8.0
eyes-selenium==5.26.0
marshmallow==2.21.0
PyRect==0.2.0
PyGetWindow==0.0.9
wsproto==1.2.0
requests==2.32.3
python-dotenv==1.0.1
packaging==24.1